+++ io_osu_beatmaps_replays/cursor.py
@@ -2,9 +2,8 @@
 
 import bpy
 from .utils import get_ms_per_frame, map_osu_to_blender
-from .geometry_nodes import create_geometry_nodes_modifier_cursor
+from .geometry_nodes import create_geometry_nodes_modifier_cursor, connect_attributes_with_drivers
 from .osu_replay_data_manager import OsuReplayDataManager
-from .exec import connect_attributes_with_drivers
 
 def create_cursor(cursor_collection, data_manager: OsuReplayDataManager):
     try:
