+++ io_osu_beatmaps_replays/slider.py
@@ -5,10 +5,9 @@ import math
 from mathutils import Vector
 from .constants import SCALE_FACTOR
 from .utils import map_osu_to_blender, get_ms_per_frame
-from .geometry_nodes import create_geometry_nodes_modifier_slider
+from .geometry_nodes import create_geometry_nodes_modifier_slider, connect_attributes_with_drivers
 from .osu_replay_data_manager import OsuReplayDataManager
 from .hitobjects import HitObject
-from .exec import connect_attributes_with_drivers
 
 
 class SliderCreator:
